empty-trash.sh - added '/usr/bin/trash-empty' to empty-trash()

+++ empty-trash.sh
@@ -7,7 +7,7 @@
 # Author       : Copyright © 2023, Richard B. Romig, Mosfanet
 # Email        : rick.romig@gmail.com | rick.romig@mymetronet.com
 # Created      : 21 Nov 2023
-# Last updated : 01 Sep 2024 Version 2.3.24245
+# Last updated : 02 Sep 2024 Version 2.5.24246
 # Comments     : Run as a user cron job.
 #              : Trash directory does not exist until a file is moved to the trash.
 #              : Tested with Debian 11/12, LMDE 6, Mint 21.x, Mint 22, MX Linux 23.1, BunsenLabs 11.
@@ -23,9 +23,13 @@ empty_trash() {
 	if [[ $(find "$trash_dir"/info -type f | wc -l) -gt 0 ]]; then
 		printf "\nTrash contents:\n---------------\n"
 		tree -a "$trash_dir"
-		find "$trash_dir"/files -mindepth 1 -type f,d -exec rm -rf {} +
-		rm -f "$trash_dir"/info/*
-		printf "\nTrash emptied.\n"
+		if [[ -x usr/bin/trash-empty ]]; then
+			/usr/bin/trash-empty
+		else
+			find "$trash_dir"/files -mindepth 1 -type f,d -exec rm -rf {} +
+			rm -f "$trash_dir"/info/*
+			printf "\nTrash emptied.\n"
+		fi
 		tree -a "$trash_dir"
 	else
 		printf "\nNo trash to empty.\n"
